pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "grain"
version = "0.2.14"
description = "Grain: A library for loading and transforming data for ML training."
keywords = []
authors = [
    {name = "Grain team", email = "grain-dev@google.com"},
]
dependencies = [
    "absl-py",
    "array-record>=0.8.1; sys_platform != 'win32'",
    "cloudpickle",
    "etils[epath,epy]",
    "numpy",
    "protobuf>=5.28.3",
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
]

[project.optional-dependencies]
parquet = [
    "pyarrow",
]

[project.urls]
homepage = "https://github.com/google/grain"

[tool.setuptools.packages.find]
include = ["grain*"]

[tool.setuptools.package-data]
"*" = ["*.so", "*.pyd", "testdata/*.array_record-*", "testdata/*.tfrecord"]